arranging both front end and backend functionalities for authorization

+++ client/src/components/ExpenseDetails.js
@@ -1,5 +1,5 @@
 import { formatDistanceToNow } from "date-fns";
-import { set } from "date-fns/esm";
+import { useAuthContext } from "../hooks/useAuthContext";
 import React, { useState } from "react";
 import LoadingPage from "./Loading";
 import AlertBox from "./AlertBox";
@@ -10,15 +10,21 @@ const ExpenseDetails = ({result, cancelClick}) => {
     const [success, setSuccess] = useState(false)
     const [error, setError] = useState(false)
     const [message, setMessage] = useState(null)
+    const { user } = useAuthContext()
 
     const handleApproval = async (id) =>{
 
         setIsPending(true)
 
+        if(!user){
+            return
+        }
+
         let response = await fetch(`https://expesetracker.herokuapp.com/api/expense/${id}/approve`, {
             method: 'PATCH',
             headers: {
-                'Content-Type': 'application/json'
+                'Content-Type': 'application/json',
+                'Authorization': `Bearer ${user.token}`
             }
         })
 
@@ -42,10 +48,16 @@ const ExpenseDetails = ({result, cancelClick}) => {
 
         setIsPending(true)
 
+        if(!user){
+            return
+        }
+
+
         let response = await fetch(`https://expesetracker.herokuapp.com/api/expense/${id}`, {
             method: 'PATCH',
             headers: {
-                'Content-Type': 'application/json'
+                'Content-Type': 'application/json',
+                'Authorization': `Bearer ${user.token}`
             }
         })
 
@@ -69,10 +81,15 @@ const ExpenseDetails = ({result, cancelClick}) => {
 
         setIsPending(true)
 
+        if(!user){
+            return
+        }
+
         const response = await fetch(`https://expesetracker.herokuapp.com/api/expense/${id}`, {
             method: 'DELETE',
             headers:{
-                'Content-Type': 'application/json'
+                'Content-Type': 'application/json',
+                'Authorization': `Bearer ${user.token}`
             }
         })
